refactor(session): extract ui/ sub-package — move 14 UI bridge files
Move 5 source + 9 test files from internal/agent/session/ to
internal/agent/session/ui/. Update package declaration to 'package ui'
(panic_test.go uses external test package 'package ui_test').
Drop 'ui_bridge_' prefix from filenames (package name 'ui' provides namespace).

No type renames yet — UIBridge is still UIBridge.
session_manager.go will be fixed in Task B3.
