+++ internal/agent/session/ui/bridge_test.go
@@ -1,7 +1,7 @@
 // Copyright (c) 2026 gosharplite@gmail.com
 // SPDX-License-Identifier: MIT
 
-package session
+package ui
 
 import (
 	"context"
